+++ packages/alpha/src/types.ts
@@ -1,6 +1,11 @@
 // Copyright 2020 Cognite AS
 
-import type { DateRange, SinglePatch, SortOrder } from '@cognite/sdk';
+import type {
+  AggregateResponse,
+  DateRange,
+  SinglePatch,
+  SortOrder,
+} from '@cognite/sdk';
 import type {
   CogniteExternalId,
   CogniteInternalId,
@@ -84,6 +89,14 @@ export interface SimulatorCreate {
   unitQuantities?: SimulatorUnitQuantity[];
 }
 
+export type SimulatorIntegrationConnectorStatus = 'IDLE' | 'RUNNING_SIMULATION';
+
+export const SimulatorIntegrationConnectorStatus = {
+  IDLE: 'IDLE' as SimulatorIntegrationConnectorStatus,
+  RUNNING_SIMULATION:
+    'RUNNING_SIMULATION' as SimulatorIntegrationConnectorStatus,
+};
+
 export interface SimulatorIntegration {
   id: CogniteInternalId;
   externalId: CogniteExternalId;
@@ -94,7 +107,7 @@ export interface SimulatorIntegration {
   simulatorVersion: string;
   licenseStatus?: string;
   licenseLastCheckedTime?: Date;
-  connectorStatus?: string;
+  connectorStatus?: SimulatorIntegrationConnectorStatus;
   connectorStatusUpdatedTime?: Date;
   logId: CogniteInternalId;
   createdTime: Date;
@@ -318,6 +331,19 @@ export interface SimulatorModelCreate {
   description?: string;
   dataSetId: CogniteInternalId;
   type?: string;
+}
+
+/**
+ * Response from models aggregate endpoint
+ */
+export type SimulatorModelAggregate = AggregateResponse;
+
+export interface SimulatorModelAggregateQuery {
+  /**
+   * Filter on models with strict matching.
+   */
+  filter?: SimulatorModelFilter;
+  aggregate: 'count';
 }
 
 export interface SimulatorModelFilter {